split e2e tests for unstable_after into app and pages, run app tests for both nodejs and edge runtimes
