+++ test/e2e/app-dir/next-after-app/app/invalid-in-client/page.js
@@ -1,0 +1,13 @@
+// 'use client'
+
+import { unstable_after as after } from 'next/server'
+import { cliLog } from '../../utils/log'
+
+export const dynamic = 'force-dynamic'
+
+export default function Page() {
+  after(() => {
+    cliLog({ source: '[page] /invalid-in-client' })
+  })
+  return <div>Invalid: Client page with after()</div>
+}
