+++ test/e2e/app-dir/next-after-app/middleware.js
@@ -1,0 +1,24 @@
+import { cookies } from 'next/headers'
+import { NextResponse, unstable_after as after } from 'next/server'
+import { cliLog } from './utils/log'
+
+export function middleware(
+  /** @type {import ('next/server').NextRequest} */ request
+) {
+  const url = new URL(request.url)
+  if (url.pathname.startsWith('/middleware/redirect-source')) {
+    const requestId = url.searchParams.get('requestId')
+    after(() => {
+      cliLog({
+        source: '[middleware] /middleware/redirect-source',
+        requestId,
+        cookies: { testCookie: cookies().get('testCookie')?.value },
+      })
+    })
+    return NextResponse.redirect(new URL('/middleware/redirect', request.url))
+  }
+}
+
+export const config = {
+  matcher: '/middleware/:path*',
+}
