+++ test/e2e/app-dir/next-after-app/next.config.js
