+++ test/e2e/app-dir/next-after-pages/next.config.js
@@ -1,0 +1,7 @@
+/** @type {import('next').NextConfig} */
+module.exports = {
+  experimental: {
+    after: true,
+    testProxy: true,
+  },
+}
