+++ test/e2e/app-dir/next-after-pages/pages/middleware/redirect/index.js
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <div>Redirect</div>
+}
