+++ test/e2e/app-dir/next-after-pages/pages/pages-dir/invalid-in-page.js
@@ -1,0 +1,15 @@
+import { unstable_after as after } from 'next/server'
+import { cliLog } from '../../utils/log'
+
+export async function getServerSideProps() {
+  return { props: {} }
+}
+
+export default function Page() {
+  after(() => {
+    cliLog({
+      source: '[pages-dir] /pages-dir/invalid-in-page',
+    })
+  })
+  return <div>Invalid in Pages router Page component</div>
+}
