--- test/e2e/app-dir/next-after/.gitignore
@@ -1,1 +1,0 @@
-after-output.jsonl
