--- test/e2e/app-dir/next-after/_pages/pages-dir/invalid-in-page.js
@@ -1,15 +1,0 @@
-import { unstable_after as after } from 'next/server'
-import { cliLog } from '../../utils/log-cli'
-
-export async function getServerSideProps() {
-  return { props: {} }
-}
-
-export default function Page() {
-  after(() => {
-    cliLog({
-      source: '[pages-dir] /pages-dir/invalid-in-page',
-    })
-  })
-  return <div>Invalid in Pages router Page component</div>
-}
