--- test/e2e/app-dir/next-after/app/interrupted/calls-redirect/page.js
@@ -1,12 +1,0 @@
-import { redirect } from 'next/navigation'
-import { unstable_after as after } from 'next/server'
-import { persistentLog } from '../../../utils/log'
-
-export default function Page() {
-  after(() => {
-    persistentLog({
-      source: '[page] /interrupted/calls-redirect',
-    })
-  })
-  redirect('/interrupted/redirect-target')
-}
