--- test/e2e/app-dir/next-after/app/layout.js
@@ -1,10 +1,0 @@
-export default function AppLayout({ children }) {
-  return (
-    <html>
-      <head>
-        <title>after</title>
-      </head>
-      <body>{children}</body>
-    </html>
-  )
-}
