--- test/e2e/app-dir/next-after/app/route/route.js
@@ -1,13 +1,0 @@
-import { unstable_after as after } from 'next/server'
-import { persistentLog } from '../../utils/log'
-
-export const dynamic = 'force-dynamic'
-
-export async function GET() {
-  const data = { message: 'Hello, world!' }
-  after(() => {
-    persistentLog({ source: '[route handler] /route' })
-  })
-
-  return Response.json({ data })
-}
